Set up the app with a test environment
We need to test this app, so this adds a test config class and changes
the app to use the application factory pattern so that we can reuse the
code in `main.py` more easily for the test Flask app.

This also sets up a single blueprint, main. Without this, all of the
routes would need to be defined inside the `create_app` function, so the
blueprint makes things easier to read.